Correção no LRU e organização

+++ include/utils.h
@@ -4,8 +4,8 @@
 #include <stddef.h>
 #include <sys/types.h>
 
-ssize_t safe_read(int fd, void *buf, size_t count);
-ssize_t safe_write(int fd, const void *buf, size_t count);
+int safe_read(int fd, void *buf, size_t count);
+int safe_write(int fd, const void *buf, size_t count);
 
 void   *xmalloc(size_t n);
 void   *xcalloc(size_t nmemb, size_t size);
